+++ tests/_conftest/__init__.py
@@ -1,3 +1,4 @@
 from .compiled_account import read_compiled_account_contract
+from .devnet_account import DevnetAccount
 from .devnet_account_preparator import DevnetAccountPreparator
 from .faucet_contract import FaucetContract
